Forgot to refactor some stuff from the previous commit

+++ src/main/java/org/cubyte/trafficsignalizer/signal/stress/CarCountStressFunction.java
@@ -28,8 +28,8 @@ public class CarCountStressFunction implements StressFunction {
 
     @Override
     public double calculateStress(Network network, Signal signal, SignalSystem system, double timeSeconds) {
-        int actual = allKnowingTrafficTracker.carCountAt(signal.getLinkId());
-        int predicted = trafficTracker.carCountAt(signal.getLinkId());
+        double actual = allKnowingTrafficTracker.carCountAt(signal.getLinkId());
+        double predicted = trafficTracker.carCountAt(signal.getLinkId());
 
         if (!signalizerParams.learn && predicted != 0 && actual != 0) {
             textWriter.put("signal_prediction_" + signal.getId(), predicted + "/" + actual, network.getLinks().get(signal.getLinkId()).getToNode().getCoord());
