+++ src/main/java/org/cubyte/trafficsignalizer/tracker/TrafficTracker.java
@@ -4,5 +4,5 @@ import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 
 public interface TrafficTracker {
-    int carCountAt(Id<Link> link);
+    double carCountAt(Id<Link> link);
 }
